+++ src/services/LoginService.ts
@@ -1,36 +1,36 @@
 import axios, { AxiosInstance } from "axios";
 
 export class LoginService {
-    private API_URL = process.env.NEXT_PUBLIC_API_URL ?? "/api";
+  private API_URL = process.env.NEXT_PUBLIC_API_URL ?? "/api";
 
-    private api: AxiosInstance;
-    
-    constructor() {
-        this.api = axios.create({ 
-            baseURL: `${this.API_URL}/auth`,
-            timeout: 5000 
-        })
-    }
+  private api: AxiosInstance;
 
-    login(username: string, password: string, callback: (() => void), callbackError?: ((errorMessage: string) => void)) {
-        var formData = new URLSearchParams();
-        formData.append("username", username);
-        formData.append("password", password);
+  constructor() {
+    this.api = axios.create({
+      baseURL: `${this.API_URL}/auth`,
+      timeout: 5000
+    })
+  }
 
-        this.api.post(`/login`, formData, {auth: {username, password}}).then(callback)
-        .catch((error) => {
-            if (!callbackError)
-                return;
+  login(username: string, password: string, callback: (() => void), callbackError?: ((errorMessage: string) => void)) {
+    var formData = new URLSearchParams();
+    formData.append("username", username);
+    formData.append("password", password);
 
-            let message = 'Erro interno ao processar requisição';
-            try {
-                let json = JSON.parse(error.request.response);
-                if (json.message)
-                    message = json.message;
-            } catch {
-            }
+    this.api.post(`/login`, formData, { auth: { username, password } }).then(callback)
+      .catch((error) => {
+        if (!callbackError)
+          return;
 
-            callbackError(message);
-        })
-    }
+        let message = 'Erro interno ao processar requisição';
+        try {
+          let json = JSON.parse(error.request.response);
+          if (json.message)
+            message = json.message;
+        } catch {
+        }
+
+        callbackError(message);
+      })
+  }
 }
