+++ src/styles/app.css
@@ -1,15 +1,18 @@
 a {
-    color: white;
-    text-decoration: none;
+  color: white;
+  text-decoration: none;
 }
-html, body {
-    background: linear-gradient(45deg, rgb(23, 97, 139), rgb(41, 175, 253));
-    background-repeat: no-repeat;
-    background-size: cover;
-    color: white;
-}
+
+html,
 body {
-    background-image: url('/img/wallpaper.jpg');
-    background-size: cover;
-    background-position: center;
+  background: linear-gradient(45deg, rgb(23, 97, 139), rgb(41, 175, 253));
+  background-repeat: no-repeat;
+  background-size: cover;
+  color: white;
+}
+
+body {
+  background-image: url('/img/wallpaper.jpg');
+  background-size: cover;
+  background-position: center;
 }
